FAMS3-1100
ICBC Web Adapter needs to be deployed on fams3 openshift dev/test. In
order to accomplish this, the project Jenkins file needed to be edited
to include the build process of the adapter as well as deploying the
build image to dev and test. Template files of the web adapter build and
deployment config were also added.